Time: 3 ms (91.44%) | Memory: 95.2 MB (48.61%) - LeetSync

+++ 2676-find-the-score-of-all-prefixes-of-an-array/find-the-score-of-all-prefixes-of-an-array.java
@@ -5,7 +5,7 @@ class Solution {
         int max = nums[0];
         for(int i = 0; i<n ; i++){
             max = Math.max(max, nums[i]);
-            arr[i] = nums[i] + max;
+            arr[i] = (long)nums[i] + (long)max;
         }
         for(int i = 1 ; i<n ; i++){
             arr[i] += arr[i-1];
